abstracted map and controller
